Refactor type handling in concepts
- Introduce `unwrap` and `unwrap_impl` for type extraction
- Modify `CleanT` in `ContainerTraits` to use `unwrap_t`
- Change `matrixRandom` and `vectorRandom` to use `size_t`
- Update `VectorVariant` for better clarity
- Switch `TDArray` from `std::array` to `std::vector`

+++ src/pub/operations/Builders.hpp
@@ -213,7 +213,7 @@ namespace mlinalg {
      * @return The random vector of the given size.
      */
     template <Number number, int n>
-    Vector<number, n> vectorRandom(const int size, const number min = 0, const number max = 100,
+    Vector<number, n> vectorRandom(const size_t size, const number min = 0, const number max = 100,
                                    const Seed& seed = std::nullopt) {
         if constexpr (n == Dynamic) {
             Vector<number, n> vec(size);
@@ -254,7 +254,7 @@ namespace mlinalg {
      * @return The random matrix of the given size.
      */
     template <Number number, int m, int n>
-    Matrix<number, m, n> matrixRandom(const int numRows, const int numCols, const number min = 0,
+    Matrix<number, m, n> matrixRandom(const size_t numRows, const size_t numCols, const number min = 0,
                                       const number max = 100, const Seed& seed = std::nullopt) {
         if constexpr (n == Dynamic || m == Dynamic) {
             Matrix<number, m, n> res(numRows, numCols);
